+++ scripts/update_minimal_submodules.sh
@@ -12,31 +12,6 @@ then
   source "${CURRENT_DIR}/set_vars.sh"
 fi
 
-usage() {
-  cat <<USAGE
-Usage: $0
-
-This script is used to update sources of ostis-web-platform and submodules.
-The exact behavior is configured via run arguments.
-
-Options:
-USAGE
-  exit 1
-}
-
-while [ "$1" != "" ];
-do
-  case $1 in
-    "--help" | "-h" )
-      usage
-      ;;
-    * )
-      usage
-      ;;
-  esac
-  shift 1
-done
-
 info "Update submodules"
 
 cd "${PLATFORM_PATH}" && git submodule update --init --recursive
@@ -44,7 +19,5 @@ cd "${PLATFORM_PATH}" && git submodule update --init --recursive
 "${SUBMODULE_SCRIPTS_DIR}/update_sc_machine.sh"
 "${SUBMODULE_SCRIPTS_DIR}/update_scp_machine.sh"
 "${SUBMODULE_SCRIPTS_DIR}/update_sc_component_manager.sh"
-"${SUBMODULE_SCRIPTS_DIR}/update_sc_web.sh"
-"${SUBMODULE_SCRIPTS_DIR}/update_ims_ostis_kb.sh"
 
 info "Submodules is updated successfully"